requirements.txt
asgiref==1.1.2
attrs==17.2.0
autobahn==17.10.1
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
Django==1.11.7
django-cors-headers==2.1.0
django-filter==1.1.0
django-prettyjson==0.3.0
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
hyperlink==17.3.1
incremental==17.5.0
Markdown==2.6.9
psycopg2==2.7.3.2
PyJWT==1.5.3
pytz==2017.3
six==1.11.0
standardjson==0.3.1
Twisted==17.9.0
txaio==2.8.2
zope.interface==4.4.3
